+++ src/zimscraperlib/imaging.py
@@ -10,6 +10,9 @@ import colorthief
 from resizeimage import resizeimage
 
 
+ALPHA_NOT_SUPPORTED = ["JPEG", "BMP", "EPS", "PCX"]
+
+
 def get_colors(image_path, use_palette=True):
     """ (main, secondary) HTML color codes from an image path """
 
@@ -40,22 +43,63 @@ def get_colors(image_path, use_palette=True):
     return rgb_to_hex(mr, mg, mb), rgb_to_hex(sr, sg, sb)
 
 
-def resize_image(fpath, width, height=None, to=None, method="width"):
+def save_image(image, dst, fmt, **params):
+    """ saves an image with default args and overrides them if params are given
+        params: PIL params (https://pillow.readthedocs.io/en/stable/handbook/image-file-formats.html) """
+    args = {"JPEG": {"quality": 100}, "PNG": {}}.get(fmt, {})
+    args.update(params or {})
+    image.save(dst, fmt, **args)
+
+
+def resize_image(
+    fpath, width, height=None, to=None, method="width", allow_upscaling=True, **params,
+):
     """ resize an image file (dimensions)
 
-        methods: width, height, cover """
-    with open(str(fpath), "rb") as fp:
-        with PIL.Image.open(fp) as image:
-            if method == "width":
-                resized = resizeimage.resize(method, image, width)
-            elif method == "height":
-                resized = resizeimage.resize(method, image, height)
-            else:
-                resized = resizeimage.resize(method, image, [width, height])
-    kwargs = {"JPEG": {"quality": 100}, "PNG": {}}
-    resized.save(
-        str(to) if to is not None else fpath, image.format, **kwargs.get(image.format)
-    )
+        methods: width, height, cover
+        allow upscaling: upscale image preserving aspect ratio if required before resizing """
+    with PIL.Image.open(fpath) as image:
+        # preserve image format as resize() does not transmit it into new object
+        image_format = image.format
+        image_mode = image.mode
+
+        # upscale if required preserving the aspect ratio
+        if allow_upscaling:
+            height_width_ratio = float(image.size[1]) / float(image.size[0])
+            if image.size[0] < width:
+                image = image.resize((width, int(width * height_width_ratio)))
+            if height and image.size[1] < height:
+                image = image.resize((int(height / height_width_ratio), height))
+
+        # resize using the requested method
+        if method == "width":
+            resized = resizeimage.resize(method, image, width)
+        elif method == "height":
+            resized = resizeimage.resize(method, image, height)
+        else:
+            resized = resizeimage.resize(method, image, [width, height])
+
+    # remove alpha layer if not supported and added during resizing
+    if resized.mode == "RGBA" and image_format in ALPHA_NOT_SUPPORTED:
+        resized = resized.convert(image_mode)
+    # save the image
+    save_image(resized, str(to) if to is not None else fpath, image_format, **params)
+
+
+def convert_image(src, dst, target_format, colorspace=None, **params):
+    """ convert an image file from one format to another
+
+        colorspace: RGB, ARGB, CMYK (and other PIL colorspaces)
+        target_format: JPEG, PNG, BMP (and other PIL formats) """
+    with PIL.Image.open(src) as image:
+        dst_image = image
+        if (
+            image.mode == "RGBA" and target_format in ALPHA_NOT_SUPPORTED
+        ) or colorspace:
+            dst_image = (
+                image.convert("RGB") if not colorspace else image.convert(colorspace)
+            )
+    save_image(dst_image, dst, target_format, **params)
 
 
 def is_hex_color(text):