+++ tests/imaging/test_imaging.py
@@ -3,10 +3,19 @@
 # vim: ai ts=4 sts=4 et sw=4 nu
 
 import pytest
+import pathlib
 
 from PIL import Image
+from resizeimage.imageexceptions import ImageSizeError
 
-from zimscraperlib.imaging import get_colors, is_hex_color, resize_image, create_favicon
+from zimscraperlib.imaging import (
+    get_colors,
+    is_hex_color,
+    resize_image,
+    create_favicon,
+    convert_image,
+    save_image,
+)
 
 
 def get_image_size(fpath):
@@ -58,13 +67,26 @@ def test_colors_jpg_palette(jpg_image):
 
 
 @pytest.mark.parametrize(
+    "fmt,params", [("png", None), ("jpg", {"quality": 50})],
+)
+def test_save_image(png_image, jpg_image, tmp_path, fmt, params):
+    src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
+    img = Image.open(src)
+    if params:
+        save_image(img, dst, "JPEG" if fmt == "jpg" else fmt, **params)
+    else:
+        save_image(img, dst, "JPEG" if fmt == "jpg" else fmt)
+    assert pathlib.Path(dst).exists()
+
+
+@pytest.mark.parametrize(
     "fmt", ["png", "jpg"],
 )
 def test_resize_thumbnail(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 100, 50
-    resize_image(png_image, width, height, to=dst, method="thumbnail")
+    resize_image(src, width, height, to=dst, method="thumbnail")
     tw, th = get_image_size(dst)
     assert tw <= width
     assert th <= height
@@ -77,8 +99,8 @@ def test_resize_width(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 100, 50
-    resize_image(png_image, width, height, to=dst, method="width")
-    tw, th = get_image_size(dst)
+    resize_image(src, width, height, to=dst, method="width")
+    tw, _ = get_image_size(dst)
     assert tw == width
 
 
@@ -89,8 +111,8 @@ def test_resize_height(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 100, 50
-    resize_image(png_image, width, height, to=dst, method="height")
-    tw, th = get_image_size(dst)
+    resize_image(src, width, height, to=dst, method="height")
+    _, th = get_image_size(dst)
     assert th == height
 
 
@@ -101,10 +123,10 @@ def test_resize_crop(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 5, 50
-    resize_image(png_image, width, height, to=dst, method="crop")
+    resize_image(src, width, height, to=dst, method="crop")
     tw, th = get_image_size(dst)
-    assert tw <= width
-    assert th <= height
+    assert tw == width
+    assert th == height
 
 
 @pytest.mark.parametrize(
@@ -114,10 +136,10 @@ def test_resize_cover(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 5, 50
-    resize_image(png_image, width, height, to=dst, method="cover")
+    resize_image(src, width, height, to=dst, method="cover")
     tw, th = get_image_size(dst)
-    assert tw <= width
-    assert th <= height
+    assert tw == width
+    assert th == height
 
 
 @pytest.mark.parametrize(
@@ -127,10 +149,50 @@ def test_resize_contain(png_image, jpg_image, tmp_path, fmt):
     src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
 
     width, height = 5, 50
-    resize_image(png_image, width, height, to=dst, method="contain")
+    resize_image(src, width, height, to=dst, method="contain")
     tw, th = get_image_size(dst)
     assert tw <= width
     assert th <= height
+
+
+@pytest.mark.parametrize(
+    "fmt", ["png", "jpg"],
+)
+def test_resize_upscale(png_image, jpg_image, tmp_path, fmt):
+    src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
+
+    width, height = 500, 1000
+    resize_image(src, width, height, to=dst, method="cover")
+    tw, th = get_image_size(dst)
+    assert tw == width
+    assert th == height
+
+
+@pytest.mark.parametrize(
+    "fmt", ["png", "jpg"],
+)
+def test_resize_small_image_error(png_image, jpg_image, tmp_path, fmt):
+    src, dst = get_src_dst(png_image, jpg_image, tmp_path, fmt)
+
+    width, height = 500, 1000
+    with pytest.raises(ImageSizeError):
+        resize_image(src, width, height, to=dst, method="cover", allow_upscaling=False)
+
+
+@pytest.mark.parametrize(
+    "src_fmt,dst_fmt,colorspace",
+    [("png", "JPEG", "RGB"), ("png", "BMP", None), ("jpg", "JPEG", "CMYK")],
+)
+def test_change_image_format(
+    png_image, jpg_image, tmp_path, src_fmt, dst_fmt, colorspace
+):
+    src, _ = get_src_dst(png_image, jpg_image, tmp_path, src_fmt)
+    dst = tmp_path / f"out.{dst_fmt.lower()}"
+    convert_image(src, dst, dst_fmt, colorspace=colorspace)
+    dst_image = Image.open(dst)
+    if colorspace:
+        assert dst_image.mode == colorspace
+    assert dst_image.format == dst_fmt
 
 
 @pytest.mark.parametrize(